text for whitelist/blacklist failure when crosshair points to an item Fix failure to save most recent options to INI file Fix compiler warning errors where practical - found and fixed some bugs in the process. Use /W4 going forward

+++ src/FormHelpers/ArmorHelper.cpp
@@ -33,7 +33,7 @@ uint32_t TESObjectARMOHelper::GetGoldValue() const
 	RE::EnchantmentItem* ench = TESFormHelper(m_armor, INIFile::SecondaryType::itemObjects).GetEnchantment();
 	if (!ench)
 	{
-		return m_armor->value;
+		return static_cast<uint32_t>(m_armor->value);
 	}
 
 	double costPP = 0.0;
